backend/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-CORS==4.0.0
Flask-JWT-Extended==4.5.2
Flask-SocketIO==5.3.4
python-socketio==5.9.0
Werkzeug==2.3.7
PyJWT==2.8.0
python-dotenv==1.0.0
requests==2.31.0